# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen warehouse (WMS).",
        "title": "Warehouse Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@anda.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/warehouses": {
            "get": {
                "description": "Mengambil daftar semua warehouse",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Get Semua Warehouse",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Warehouse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Membuat warehouse baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Buat Warehouse Baru",
                "parameters": [
                    {
                        "description": "Data Warehouse Baru",
                        "name": "warehouse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateWarehouse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Warehouse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/warehouses/{id}": {
            "get": {
                "description": "Mengambil satu data warehouse berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Get Warehouse Berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Warehouse"
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Warehouse tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data warehouse berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Hapus Warehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Warehouse tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Memperbarui data warehouse (bisa sebagian) berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouses"
                ],
                "summary": "Update Warehouse (Parsial)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update warehouse",
                        "name": "warehouse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateWarehouse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID atau data JSON tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Warehouse tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Inventory": {
            "type": "object",
            "properties": {
                "code_product": {
                    "type": "string"
                },
                "code_warehouse": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_size": {
                    "type": "integer"
                },
                "product": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Product"
                        }
                    ]
                },
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "$ref": "#/definitions/models.Size"
                },
                "warehouse": {
                    "$ref": "#/definitions/models.Warehouse"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ]
                },
                "description_product": {
                    "description": "Diasumsikan bisa null",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_category": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "product_code": {
                    "type": "string"
                },
                "product_details": {
                    "description": "Relasi (Has Many)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductDetail"
                    }
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
        "models.ProductDetail": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "code_product": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_size": {
                    "type": "integer"
                },
                "product": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Product"
                        }
                    ]
                },
                "size": {
                    "$ref": "#/definitions/models.Size"
                }
            }
        },
...
        "models.Size": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Warehouse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "inventories": {
                    "description": "Relasi (Has Many)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Inventory"
                    }
                },
                "location_description": {
                    "type": "string"
                },
                "warehouse_code": {
                    "type": "string"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateWarehouse": {
            "type": "object",
            "required": [
                "warehouse_code",
                "warehouse_name"
            ],
            "properties": {
                "location_description": {
                    "type": "string"
                },
                "warehouse_code": {
                    "type": "string"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "request.UpdateWarehouse": {
            "type": "object",
            "properties": {
                "location_description": {
                    "type": "string"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
...
    }
}